+++ ndc-test/src/test_cases/schema.rs
@@ -25,7 +25,23 @@ pub async fn validate_schema<R: Reporter>(
     let _ = test!("scalar_types", reporter, async {
         for (type_name, scalar_type) in &schema.scalar_types {
             for aggregate_function in scalar_type.aggregate_functions.values() {
-                validate_type(schema, &aggregate_function.result_type)?;
+                match aggregate_function {
+                    models::AggregateFunctionDefinition::Sum { result_type } => {
+                        let Some(scalar_type) = schema.scalar_types.get(result_type) else {
+                            return Err(Error::NamedTypeIsNotDefined(result_type.inner().clone()));
+                        };
+                        let (models::TypeRepresentation::Int64
+                        | models::TypeRepresentation::Float64) = scalar_type.representation
+                        else {
+                            return Err(Error::InvalidTypeRepresentation(result_type.clone()));
+                        };
+                        Ok(())
+                    }
+                    models::AggregateFunctionDefinition::Custom { result_type } => {
+                        validate_type(schema, result_type)
+                    }
+                    _ => Ok(()),
+                }?;
             }
 
             let mut has_equality = false;